+++ internal/tun/setup.go
@@ -6,7 +6,7 @@ import (
 	"github.com/ooni/minivpn/internal/model"
 	"github.com/ooni/minivpn/internal/networkio"
 	"github.com/ooni/minivpn/internal/packetmuxer"
-	"github.com/ooni/minivpn/internal/reliable"
+	"github.com/ooni/minivpn/internal/reliabletransport"
 	"github.com/ooni/minivpn/internal/runtimex"
 	"github.com/ooni/minivpn/internal/session"
 	"github.com/ooni/minivpn/internal/tlsstate"
@@ -68,8 +68,8 @@ func startWorkers(logger model.Logger, sessionManager *session.Manager,
 	connectChannel(datach.DataPacketUp, &muxer.DataPacketUp)
 	connectChannel(muxer.PacketDown, &datach.DataPacketDown)
 
-	// create the reliable service.
-	rel := &reliable.Service{
+	// create the reliabletransport service.
+	rel := &reliabletransport.Service{
 		PacketDownBottom: nil, // ok
 		PacketDownTop:    make(chan *model.Packet),
 		PacketUpBottom:   make(chan *model.Packet),
